Implemented terminator symbol handling; added tests for testing encoding/decoding; improved other tests; Huffman coding is fully working and integrated now.

+++ test/test_tools/generate_mock_contents.h
@@ -9,8 +9,8 @@ namespace squeeze::testing::tools {
 
 struct MockContentsGeneratorParams {
     std::size_t size_limit;
-    int min_nr_reps;
-    int max_nr_reps;
+    unsigned long min_nr_reps;
+    unsigned long max_nr_reps;
     unsigned char noise_probability = 25;
     enum {
         NoneFlags = 0,
